+++ apps/consent-ui/src/components/common/PageLayout/PageLayout.module.scss
